+++ graphene_tornado/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '2.1.1'
+__version__ = '2.4.0'
 
 __all__ = [
     '__version__'