+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+# 2.4.0
+
+* Adding some tooling for building observability tools ported from apollo-tooling
+
+# 2.3.0
+
+* Fix for field resolvers in extension framework
+
+# 2.2.0
+
+* Added experimental extensions framework
+
 ## 2.1.0
 
 * Upgraded to Graphene 2.1